+++ build.sh
@@ -38,7 +38,7 @@ cat <<EOF > "$APP_DIR/Contents/Info.plist"
     <key>CFBundleShortVersionString</key>
     <string>1.6.0</string>
     <key>CFBundleVersion</key>
-    <string>7</string>
+    <string>8</string>
     <key>LSMinimumSystemVersion</key>
     <string>13.0</string>
     <key>LSUIElement</key>